+++ tests/TestCase/Console/ConsoleApplicationTest.php
@@ -109,7 +109,7 @@ class ConsoleApplicationTest extends \PHPUnit\Framework\TestCase
         $consoleApplication->description(['Cache enabler and disabler']);
         $consoleApplication->addCommand('enable', CacheEnableCommand::class);
         $consoleApplication->addCommand('disable', CacheDisableCommand::class);
-        $this->assertTrue($consoleApplication->run([]));
+        $this->assertEquals(Command::SUCCESS, $consoleApplication->run([]));
         $this->assertEquals('147bd05c7164f8418d58dd1199546a02', md5($this->output->read()));
     }
 
@@ -117,7 +117,7 @@ class ConsoleApplicationTest extends \PHPUnit\Framework\TestCase
     {
         $consoleApplication = $this->consoleApplication();
         $consoleApplication->addCommand('single', FooCommand::class);
-        $this->assertFalse($consoleApplication->run());
+        $this->assertEquals(Command::ERROR, $consoleApplication->run());
     }
 
     public function testSingleCommandApp()
@@ -125,7 +125,7 @@ class ConsoleApplicationTest extends \PHPUnit\Framework\TestCase
         $consoleApplication = $this->consoleApplication();
         $consoleApplication->name('cache');
         $consoleApplication->addCommand('enable', CacheEnableCommand::class);
-        $this->assertTrue($consoleApplication->run([]));
+        $this->assertEquals(Command::SUCCESS, $consoleApplication->run([]));
         $this->assertStringContainsString('OK Cache enabled', $this->output->read());
     }
 
@@ -137,7 +137,7 @@ class ConsoleApplicationTest extends \PHPUnit\Framework\TestCase
         $consoleApplication->addCommand('disable', CacheDisableCommand::class);
         $consoleApplication->addCommand('delete', CacheDeleteCommand::class);
         
-        $this->assertFalse($consoleApplication->run(['delete']));
+        $this->assertEquals(Command::ERROR, $consoleApplication->run(['delete']));
 
         $this->assertStringContainsString('Missing required argument `key`', $this->output->read());
     }
@@ -148,7 +148,7 @@ class ConsoleApplicationTest extends \PHPUnit\Framework\TestCase
         $consoleApplication->name('cache');
         $consoleApplication->addCommand('enable', CacheEnableCommand::class);
         $consoleApplication->addCommand('disable', CacheDisableCommand::class);
-        $this->assertTrue($consoleApplication->run(['enable']));
+        $this->assertEquals(Command::SUCCESS, $consoleApplication->run(['enable']));
 
         $this->assertStringContainsString('OK Cache enabled', $this->output->read());
 
@@ -163,7 +163,7 @@ class ConsoleApplicationTest extends \PHPUnit\Framework\TestCase
         $consoleApplication->name('cache');
         $consoleApplication->addCommand('enable', CacheEnableCommand::class);
         $consoleApplication->addCommand('disable', CacheDisableCommand::class);
-        $this->assertFalse($consoleApplication->run(['reset']));
+        $this->assertEquals(Command::ERROR, $consoleApplication->run(['reset']));
         $this->assertStringContainsString('Invalid command reset', $this->output->read());
     }